:pencil: feat:
Crete class only to validate CPF and private functions that calculate validations
