+++ src/Controllers/ClientController.ts
@@ -11,13 +11,13 @@ export class ClientController {
 			birthday,
 		}: { name: string; cpf: string; birthday: string } = req.body;
 
-		await this.clientService.verifyCpf(name, cpf, birthday);
+		await this.clientService.createClient(name, cpf, birthday);
 
 		return res.status(201).send({ name, cpf, birthday });
 	}
 
 	async getClientByCPF(req: Request, res: Response) {
-		const { cpf }: { name: string; cpf: string; birthday: string } = req.body;
+		const { cpf }: { cpf: string } = req.body;
 
 		const client = await this.clientService.getClient(cpf);
